# gen/http/openapi3.json
{"openapi":"3.0.3","info":{"title":"Cache Service","description":"The cache service exposes interface for working with Redis.","version":"0.0.1"},"servers":[{"url":"http://localhost:8083","description":"Cache Server"}],"paths":{"/liveness":{"get":{"tags":["health"],"summary":"Liveness health","operationId":"health#Liveness","responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"},"example":{"service":"Laborum reprehenderit rerum est et ut dolores.","status":"Consequatur porro qui est dolor a.","version":"Ad dolor."}}}}}}},"/readiness":{"get":{"tags":["health"],"summary":"Readiness health","operationId":"health#Readiness","responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"},"example":{"service":"Fugiat voluptatem vel et.","status":"Sint tempore est nam iusto.","version":"Ipsam quidem aut velit vitae est."}}}}}}},"/v1/cache":{"get":{"tags":["cache"],"summary":"Get cache","description":"Get JSON value from the cache.","operationId":"cache#Get","parameters":[{"name":"x-cache-key","in":"header","description":"Cache entry key","allowEmptyValue":true,"required":true,"schema":{"type":"string","description":"Cache entry key","example":"did:web:example.com"},"example":"did:web:example.com"},{"name":"x-cache-namespace","in":"header","description":"Cache entry namespace","allowEmptyValue":true,"schema":{"type":"string","description":"Cache entry namespace","example":"Login"},"example":"Login"},{"name":"x-cache-scope","in":"header","description":"Cache entry scope","allowEmptyValue":true,"schema":{"type":"string","description":"Cache entry scope","example":"administration"},"examples":{"default":{"summary":"default","value":"administration"},"multiple scopes":{"summary":"multiple scopes","value":"administration,user"}}},{"name":"x-cache-flatten-strategy","in":"header","description":"Flatten strategy.","allowEmptyValue":true,"schema":{"type":"string","description":"Flatten strategy.","example":"merge"},"examples":{"default":{"summary":"default","value":"merge"},"first key value only":{"summary":"first key value only","value":"first"},"last key value only":{"summary":"last key value only","value":"last"}}}],"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"example":"Velit quaerat voluptatem."},"example":"Aut enim aut cupiditate excepturi quam sunt."}}}}},"post":{"tags":["cache"],"summary":"Set cache","description":"Set a JSON value in the cache.","operationId":"cache#Set","parameters":[{"name":"x-cache-key","in":"header","description":"Cache entry key","allowEmptyValue":true,"required":true,"schema":{"type":"string","description":"Cache entry key","example":"did:web:example.com"},"example":"did:web:example.com"},{"name":"x-cache-namespace","in":"header","description":"Cache entry namespace","allowEmptyValue":true,"schema":{"type":"string","description":"Cache entry namespace","example":"Login"},"example":"Login"},{"name":"x-cache-scope","in":"header","description":"Cache entry scope","allowEmptyValue":true,"schema":{"type":"string","description":"Cache entry scope","example":"administration"},"example":"administration"},{"name":"x-cache-ttl","in":"header","description":"Cache entry TTL in seconds","allowEmptyValue":true,"schema":{"type":"integer","description":"Cache entry TTL in seconds","example":60,"format":"int64"},"example":60}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"example":"Tenetur totam itaque ad commodi omnis voluptatem."},"example":"Quo sapiente."}}},"responses":{"201":{"description":"Created response."}}}},"/v1/external/cache":{"post":{"tags":["cache"],"summary":"SetExternal cache","description":"Set an external JSON value in the cache and provide an event for the input.","operationId":"cache#SetExternal","parameters":[{"name":"x-cache-key","in":"header","description":"Cache entry key","allowEmptyValue":true,"required":true,"schema":{"type":"string","description":"Cache entry key","example":"did:web:example.com"},"example":"did:web:example.com"},{"name":"x-cache-namespace","in":"header","description":"Cache entry namespace","allowEmptyValue":true,"schema":{"type":"string","description":"Cache entry namespace","example":"Login"},"example":"Login"},{"name":"x-cache-scope","in":"header","description":"Cache entry scope","allowEmptyValue":true,"schema":{"type":"string","description":"Cache entry scope","example":"administration"},"example":"administration"},{"name":"x-cache-ttl","in":"header","description":"Cache entry TTL in seconds","allowEmptyValue":true,"schema":{"type":"integer","description":"Cache entry TTL in seconds","example":60,"format":"int64"},"example":60}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"example":"Id quidem fugit."},"example":"Aut corrupti repellendus."}}},"responses":{"200":{"description":"OK response."}}}}},"components":{"schemas":{"CacheGetRequest":{"type":"object","properties":{"key":{"type":"string","example":"Rerum et qui alias qui."},"namespace":{"type":"string","example":"Beatae commodi."},"scope":{"type":"string","example":"Fuga voluptas explicabo et libero."},"strategy":{"type":"string","example":"Illum nam ratione nisi."}},"example":{"key":"Labore vel.","namespace":"Aut illum.","scope":"Itaque vel.","strategy":"Itaque enim aut consequatur beatae ut."},"required":["key"]},"CacheSetRequest":{"type":"object","properties":{"data":{"example":"Et atque impedit nostrum perspiciatis ipsum."},"key":{"type":"string","example":"Accusamus ratione voluptatibus."},"namespace":{"type":"string","example":"Quae minus maiores nulla deleniti ipsa."},"scope":{"type":"string","example":"Quidem nihil quis tempore."},"ttl":{"type":"integer","example":4775933368413984102,"format":"int64"}},"example":{"data":"Quis doloremque iure.","key":"Reiciendis autem perferendis porro laborum.","namespace":"Dolorem aut nesciunt placeat sed.","scope":"Qui placeat sed fugit amet.","ttl":8905228881162231273},"required":["data","key"]},"HealthResponse":{"type":"object","properties":{"service":{"type":"string","description":"Service name.","example":"Ex minus nostrum eaque unde eligendi."},"status":{"type":"string","description":"Status message.","example":"Qui porro beatae porro."},"version":{"type":"string","description":"Service runtime version.","example":"Voluptatem facere commodi facilis magnam officia."}},"example":{"service":"Tempore provident laborum et perferendis.","status":"Laudantium aut tempora.","version":"Quidem omnis quia et facere."},"required":["service","status","version"]}}},"tags":[{"name":"cache","description":"Cache service allows storing and retrieving data from distributed cache."},{"name":"health","description":"Health service provides health check endpoints."}]}